requirements.txt
# Testing
pytest
pytest-selenium
allure-pytest
pytest-check

# Web & Mobile Testing
Appium-Python-Client
selenium

# Code Quality
flake8
isort
black

# Utilities
colorama
pyyaml
numpy
pyautogui
opencv-python
allure-python-commons
webdriver-manager
